seperate module into different file and test unit
