+++ modules/reverse.js
@@ -1,0 +1,1 @@
+export const reverse = (string) => string.split('').reverse().join('');
